requirements.txt
googletrans==4.0.0rc1
openai==1.43.0
Pillow==10.4.0
protobuf==5.28.0
PyYAML==6.0.1
Requests==2.32.3
transformers==4.43.3
translate==3.6.1
